+++ src/qtt/algorithms/random_telegraph_signal.py
@@ -12,8 +12,9 @@ import numpy as np
 import qcodes
 
 from qtt.utilities.tools import addPPTslide
-from qtt.algorithms.functions import double_gaussian, fit_double_gaussian, exp_function, fit_exp_decay,\
-    gaussian, refit_double_gaussian
+from qtt.algorithms.functions import double_gaussian, exp_function, fit_exp_decay, gaussian
+from qtt.algorithms.fitting import fit_double_gaussian, refit_double_gaussian
+    
 from qtt.algorithms.markov_chain import ContinuousTimeMarkovModel
 from qtt.utilities.visualization import plot_vertical_line, plot_double_gaussian_fit
